Add files via upload
A few Metallic Shaders Using similar logic to the ones I had working on Blender but through Unity itself. Includes 5 shaders plus the graphs themselves. (To create a new shader "save as" the graph with the new gradient then apply it to a new material and then use that material with an asset to apply)